back_end/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
Faker==19.3.0
Pillow==10.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
